feat(.github): update post_deploy to add lighthouse report to step summary
Updated to run post_deploy for all actions except when pull request closed. When event is `pull_request`, will add comment to the PR with report details, will add report details as a step summary to the action every time so that will have report available for when event is `push`.