chore: remove attempt for registering custom domain (#173)
<!-- markdownlint-disable first-line-heading -->

## Details

Decided to leave custom domain as a manual step for now. Keeps failing
with error `CNAME Record is invalid. Please ensure the CNAME record has
been created.`, while it has been created.

So, leaving it as a manual step in the process, fortunately only needed
once for creation of the SWA.

## Checklist

- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works

<!-- Additional details -->
<details>
<summary><strong>Expand for additional details</strong></summary>

## Related issues

<!--
A link to any related issues or bugs that the pull request
addresses, connecting the code's context with the problem it
solves.
-->
#169 

## Screenshots

<!--
If the changes are visual, including screenshots or GIFs can
help reviewers understand them more easily.
-->

## Testing instructions

<!--
Instructions on how to test the changes made in the pull
request, helping reviewers validate the code.
-->

</details>
<!-- End of Additional details -->