feat(.github): get affected projects when testing (#181)
<!-- markdownlint-disable first-line-heading -->

## Details

Moved step to check if project is affected by changes from CD workflow
into CI workflow. Checking this before any checkout in CD process allows
for quicker skip when project is not affected.

The CI workflow outputs all affected applications as a comma-separated
string. And passes this output into CD process.

CD process has a new job which checks for the project name in the new
`affected_projects` input from CI process. The build and deploy job
depends on this new job, and will be skipped if the job failed to find
the project.

### Improvements to handling affected projects:

*
[`.github/workflows/_workflow.jobs.cd.yml`](diffhunk://#diff-5d2d5523aa609da0f4aca508e5679ebb493f141329d1fd0dd7ff55b31cc840cfR21-R24):
Added a new input `affected_projects` and updated the `check_affected`
job to determine if a project is affected by the latest changes. The
`build_deploy` job now depends on the `check_affected` job and only runs
if the project is affected or `force_deploy` is true.
[[1]](diffhunk://#diff-5d2d5523aa609da0f4aca508e5679ebb493f141329d1fd0dd7ff55b31cc840cfR21-R24)
[[2]](diffhunk://#diff-5d2d5523aa609da0f4aca508e5679ebb493f141329d1fd0dd7ff55b31cc840cfR53-R78)
*
[`.github/workflows/_workflow.jobs.cd.yml`](diffhunk://#diff-5d2d5523aa609da0f4aca508e5679ebb493f141329d1fd0dd7ff55b31cc840cfL98-R138):
Removed redundant steps for checking if a project is affected within the
`build_deploy` job.
[[1]](diffhunk://#diff-5d2d5523aa609da0f4aca508e5679ebb493f141329d1fd0dd7ff55b31cc840cfL98-R138)
[[2]](diffhunk://#diff-5d2d5523aa609da0f4aca508e5679ebb493f141329d1fd0dd7ff55b31cc840cfL158)
[[3]](diffhunk://#diff-5d2d5523aa609da0f4aca508e5679ebb493f141329d1fd0dd7ff55b31cc840cfL171)
[[4]](diffhunk://#diff-5d2d5523aa609da0f4aca508e5679ebb493f141329d1fd0dd7ff55b31cc840cfL197)

### Outputs for affected projects:

*
[`.github/workflows/_workflow.jobs.ci.yml`](diffhunk://#diff-0ac98ed7fe183c1f81d4f37f880b8be8349329d6de9c3fdf8114f85ddb8c177cR21-R22):
Added outputs for affected projects in the `test_and_lint` job and
included a step to get affected applications.
[[1]](diffhunk://#diff-0ac98ed7fe183c1f81d4f37f880b8be8349329d6de9c3fdf8114f85ddb8c177cR21-R22)
[[2]](diffhunk://#diff-0ac98ed7fe183c1f81d4f37f880b8be8349329d6de9c3fdf8114f85ddb8c177cR37-R38)
[[3]](diffhunk://#diff-0ac98ed7fe183c1f81d4f37f880b8be8349329d6de9c3fdf8114f85ddb8c177cR57-R69)
*
[`.github/workflows/ci_cd.yml`](diffhunk://#diff-c80d54cb778b5196ee7b73e152870cac313eed095f65886e3f15247cdc900d1aR60):
Updated the `affected_projects` input to use the output from the
`lint_test_build` job.

## Checklist

- [x] I have made corresponding changes to the documentation
- [x] I have added tests that prove my fix is effective or that my
feature works

<!-- Additional details -->
<details>
<summary><strong>Expand for additional details</strong></summary>

## Related issues

<!--
A link to any related issues or bugs that the pull request
addresses, connecting the code's context with the problem it
solves.
-->

## Screenshots

<!--
If the changes are visual, including screenshots or GIFs can
help reviewers understand them more easily.
-->

## Testing instructions

<!--
Instructions on how to test the changes made in the pull
request, helping reviewers validate the code.
-->

</details>
<!-- End of Additional details -->