feat(bb-score): add update toast component for version notifications (#202)
<!-- markdownlint-disable first-line-heading -->

## Checklist

- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works

## Details

- Introduced a new `UpdateToastComponent` to notify users when a new
version of the application is available.
- Updated `app.component.html` to include the `app-update-toast`
component for displaying update notifications.
- Modified `app.component.ts` to import the `UpdateToastComponent`.
- Enhanced unit tests for `AppComponent` to mock the
`UpdateToastComponent` during testing.
- Created `SwUpdateService` to manage service worker updates and notify
the application of version changes.
- Added tests for `SwUpdateService` to ensure correct behavior when
updates are available and activated.

<!-- Additional details -->
<details>
<summary><strong>Expand for additional details</strong></summary>

## Related issues

<!--
A link to any related issues or bugs that the pull request
addresses, connecting the code's context with the problem it
solves.
-->

## Screenshots

<!--
If the changes are visual, including screenshots or GIFs can
help reviewers understand them more easily.
-->

## Testing instructions

<!--
Instructions on how to test the changes made in the pull
request, helping reviewers validate the code.
-->

</details>
<!-- End of Additional details -->