fix(bb-score): update snack bar for version update (#203)
<!-- markdownlint-disable first-line-heading -->

## Checklist

- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works

## Details

### Minor style fixes to include with this fix

- Refactored the GameCreateComponent to enhance the layout by replacing
the mat-card structure with a more semantic header and form layout.
- Improved styling for the form fields and button container to ensure
better responsiveness and user experience.
- Maintained existing functionality while streamlining the component's
structure for clarity and maintainability.

feat(bb-score): hide sidenav when loaded with mobile

### Actual fix

Component to be used for toast, rendered directly. Updated to only
display content when in a snackbar, and pass in needed value to know
that it is in a snackbar when opening the snackbar.

Testing of update to be done...

<!-- Additional details -->
<details>
<summary><strong>Expand for additional details</strong></summary>

## Related issues

<!--
A link to any related issues or bugs that the pull request
addresses, connecting the code's context with the problem it
solves.
-->

## Screenshots

<!--
If the changes are visual, including screenshots or GIFs can
help reviewers understand them more easily.
-->

## Testing instructions

<!--
Instructions on how to test the changes made in the pull
request, helping reviewers validate the code.
-->

</details>
<!-- End of Additional details -->