feat(bb-score): start framework for scoring (#215)
<!-- markdownlint-disable first-line-heading -->

## Checklist

- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works

## Details

Created scoring section to be used for scoring games.
Added `ScoringService` to be used to generate and append snapshots to
the game.
Updated `GameService` to be able to append snapshots.

After this, need to continue updating `ScoringService` to handle
actions, and update display components to render state correctly.