feat(bb-score) start implementing hitting for scoring (#219)
<!-- markdownlint-disable first-line-heading -->

## Checklist

- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my
feature works

## Details

<!-- Change details -->

Initial implementation of the hitting mechanics related to scoring in
the system. Further refinements and additional logic will be added in
future updates.

<!-- Additional details -->
<details>
<summary><strong>Expand for additional details</strong></summary>

## Related issues

<!--
A link to any related issues or bugs that the pull request
addresses, connecting the code's context with the problem it
solves.
-->

## Screenshots

<!--
If the changes are visual, including screenshots or GIFs can
help reviewers understand them more easily.
-->

## Testing instructions

<!--
Instructions on how to test the changes made in the pull
request, helping reviewers validate the code.
-->

</details>
<!-- End of Additional details -->